requirements.txt
kafka-python
apache-airflow

